* Diffuse transmitter can now have an internal medium. * Added whitepoint picker to GUI. * File/Save can now save scene as serialised scene graph. * Major improvements and a major bug fix for atmospheric scattering. * Network render protocol now serialises the scene, then sends individual dependencies as needed. * Added realtime updates over the network when network rendering. * Made some changes and bug fixes to MLT path tracer, to do with power heuristic weighting and participating media. * Improved node serialisation and deserialisation, added some unit tests for it. * changed IndigoVector and js::Vector to use size_t instead of unsigned int. * Dependency paths stored in the SG should now be fully absolute. * MLTpathtracer heterogenous participating media scattering is now faster and less noisy / more robust. * Added Particpating_media_scattering.pdf in code_documentation.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@1915 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/Vector.h
@@ -33,33 +33,33 @@ Vector
 ------
 Similar to std::vector, but aligns the elements.
 =====================================================================*/
-template <class T, int alignment>
+template <class T, size_t alignment>
 class Vector
 {
 public:
 	inline Vector();
-	inline Vector(unsigned int count);
-	inline Vector(unsigned int count, const T& val);
+	inline Vector(size_t count);
+	inline Vector(size_t count, const T& val);
 	inline Vector(const Vector& other);
 	inline ~Vector();
 
 	inline Vector& operator=(const Vector& other);
 
-	inline void reserve(unsigned int N); // Make sure capacity is at least N.
-	inline void resize(unsigned int N);
-	inline unsigned int capacity() const { return capacity_; }
-	inline unsigned int size() const;
+	inline void reserve(size_t N); // Make sure capacity is at least N.
+	inline void resize(size_t N);
+	inline size_t capacity() const { return capacity_; }
+	inline size_t size() const;
 	inline bool empty() const;
 	inline void clearAndFreeMem(); // Set size to zero, but also frees actual array memory.
 
-	inline void clearAndSetCapacity(unsigned int N);
+	inline void clearAndSetCapacity(size_t N);
 
 	inline void push_back(const T& t);
 	inline void pop_back();
 	inline const T& back() const;
 	inline T& back();
-	inline T& operator[](unsigned int index);
-	inline const T& operator[](unsigned int index) const;
+	inline T& operator[](size_t index);
+	inline const T& operator[](size_t index) const;
 
 	typedef T* iterator;
 	typedef const T* const_iterator;
@@ -72,15 +72,15 @@ public:
 private:
 	
 	
-	static inline void copy(const T* const src, T* dst, unsigned int num);
+	static inline void copy(const T* const src, T* dst, size_t num);
 
 	T* e;
-	unsigned int size_;
-	unsigned int capacity_;
+	size_t size_;
+	size_t capacity_;
 };
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 Vector<T, alignment>::Vector()
 :	e(0),
 	size_(0),
@@ -90,8 +90,8 @@ Vector<T, alignment>::Vector()
 }
 
 
-template <class T, int alignment>
-Vector<T, alignment>::Vector(unsigned int count)
+template <class T, size_t alignment>
+Vector<T, alignment>::Vector(size_t count)
 :	e(0),
 	size_(0),
 	capacity_(0)
@@ -103,8 +103,8 @@ Vector<T, alignment>::Vector(unsigned int count)
 }
 
 
-template <class T, int alignment>
-Vector<T, alignment>::Vector(unsigned int count, const T& val)
+template <class T, size_t alignment>
+Vector<T, alignment>::Vector(size_t count, const T& val)
 :	e(0),
 	size_(0),
 	capacity_(0)
@@ -112,12 +112,12 @@ Vector<T, alignment>::Vector(unsigned int count, const T& val)
 	assert(alignment > sizeof(T) || sizeof(T) % alignment == 0); // sizeof(T) needs to be a multiple of alignment, otherwise e[1] will be unaligned.
 
 	resize(count);
-	for(unsigned int i=0; i<count; ++i)
+	for(size_t i=0; i<count; ++i)
 		e[i] = val;
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 Vector<T, alignment>::Vector(const Vector<T, alignment>& other)
 :	e(0),
 	size_(0),
@@ -127,7 +127,7 @@ Vector<T, alignment>::Vector(const Vector<T, alignment>& other)
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 Vector<T, alignment>::~Vector()
 {
 	assert(capacity_ >= size_);
@@ -137,7 +137,7 @@ Vector<T, alignment>::~Vector()
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 Vector<T, alignment>& Vector<T, alignment>::operator=(const Vector& other)
 {
 	assert(capacity_ >= size_);
@@ -157,8 +157,8 @@ Vector<T, alignment>& Vector<T, alignment>::operator=(const Vector& other)
 }
 
 
-template <class T, int alignment>
-void Vector<T, alignment>::reserve(unsigned int n)
+template <class T, size_t alignment>
+void Vector<T, alignment>::reserve(size_t n)
 {
 	assert(capacity_ >= size_);
 
@@ -190,15 +190,15 @@ void Vector<T, alignment>::reserve(unsigned int n)
 }
 
 
-template <class T, int alignment>
-void Vector<T, alignment>::resize(unsigned int n)
+template <class T, size_t alignment>
+void Vector<T, alignment>::resize(size_t n)
 {
 	assert(capacity_ >= size_);
 
 	//reserve(n);
 	if(n > capacity_)
 	{
-		const unsigned int newcapacity = myMax(n, capacity_ + capacity_ / 2); // current size * 1.5
+		const size_t newcapacity = myMax(n, capacity_ + capacity_ / 2); // current size * 1.5
 		//const unsigned int newcapacity = ideal_newcapacity > (size_ + 1) ? ideal_newcapacity : (size_ + 1); //(size_ + 1) * 2;
 		//assert(newcapacity >= size_ + 1);
 		reserve(newcapacity);
@@ -211,21 +211,21 @@ void Vector<T, alignment>::resize(unsigned int n)
 }
 
 
-template <class T, int alignment>
-unsigned int Vector<T, alignment>::size() const
+template <class T, size_t alignment>
+size_t Vector<T, alignment>::size() const
 {
 	return size_;
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 bool Vector<T, alignment>::empty() const
 {
 	return size_ == 0;
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 void Vector<T, alignment>::clearAndFreeMem() // Set size to zero, but also frees actual array memory.
 {
 	SSE::alignedFree(e); // Free old buffer.
@@ -235,15 +235,15 @@ void Vector<T, alignment>::clearAndFreeMem() // Set size to zero, but also frees
 }
 
 
-template <class T, int alignment>
-inline void Vector<T, alignment>::clearAndSetCapacity(unsigned int N)
+template <class T, size_t alignment>
+inline void Vector<T, alignment>::clearAndSetCapacity(size_t N)
 {
 	clearAndFreeMem();
 	reserve(N);
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 void Vector<T, alignment>::push_back(const T& t)
 {
 	assert(capacity_ >= size_);
@@ -255,7 +255,7 @@ void Vector<T, alignment>::push_back(const T& t)
 		assert(newcapacity >= size_ + 1);
 		reserve(newcapacity);
 	}*/
-	const unsigned int old_size = size_;
+	const size_t old_size = size_;
 	resize(size_ + 1);
 
 	e[old_size] = t;
@@ -265,7 +265,7 @@ void Vector<T, alignment>::push_back(const T& t)
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 void Vector<T, alignment>::pop_back()
 {
 	assert(capacity_ >= size_);
@@ -279,7 +279,7 @@ void Vector<T, alignment>::pop_back()
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 const T& Vector<T, alignment>::back() const
 {
 	assert(capacity_ >= size_);
@@ -289,7 +289,7 @@ const T& Vector<T, alignment>::back() const
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 T& Vector<T, alignment>::back()
 {
 	assert(size_ >= 1);
@@ -299,19 +299,19 @@ T& Vector<T, alignment>::back()
 }
 
 
-template <class T, int alignment>
-void Vector<T, alignment>::copy(const T * const src, T* dst, unsigned int num)
+template <class T, size_t alignment>
+void Vector<T, alignment>::copy(const T * const src, T* dst, size_t num)
 {
 	assert(num == 0 || src);
 	assert(num == 0 || dst);
 
-	for(unsigned int i=0; i<num; ++i)
+	for(size_t i=0; i<num; ++i)
 		dst[i] = src[i];
 }
 
 
-template <class T, int alignment>
-T& Vector<T, alignment>::operator[](unsigned int index)
+template <class T, size_t alignment>
+T& Vector<T, alignment>::operator[](size_t index)
 {
 	assert(capacity_ >= size_);
 	assert(index < size_);
@@ -320,8 +320,8 @@ T& Vector<T, alignment>::operator[](unsigned int index)
 }
 
 
-template <class T, int alignment>
-const T& Vector<T, alignment>::operator[](unsigned int index) const
+template <class T, size_t alignment>
+const T& Vector<T, alignment>::operator[](size_t index) const
 {
 	assert(index < size_);
 	assert(capacity_ >= size_);
@@ -330,28 +330,28 @@ const T& Vector<T, alignment>::operator[](unsigned int index) const
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 typename Vector<T, alignment>::iterator Vector<T, alignment>::begin()
 {
 	return e;
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 typename Vector<T, alignment>::iterator Vector<T, alignment>::end()
 {
 	return e + size_;
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 typename Vector<T, alignment>::const_iterator Vector<T, alignment>::begin() const
 {
 	return e;
 }
 
 
-template <class T, int alignment>
+template <class T, size_t alignment>
 typename Vector<T, alignment>::const_iterator Vector<T, alignment>::end() const
 {
 	return e + size_;
